requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==6.1.4
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
gunicorn==20.1.0
google-cloud-storage
google-auth
